+++ mod-di-converter-storage-server/src/main/java/org/folio/rest/impl/DataImportProfilesImpl.java
@@ -83,6 +83,7 @@ public class DataImportProfilesImpl implements DataImportProfiles {
   private static final String INVALID_ACTION_TYPE_LINKED_ACTION_PROFILE_TO_MAPPING_PROFILE = "Unable to complete requested change. " +
     "MARC Update Action profiles can only be linked with MARC Update Mapping profiles and MARC Modify Action profiles can only be linked with MARC Modify Mapping profiles. " +
     "Please ensure your Action and Mapping profiles are of like types and try again.";
+  private static final String INVALID_ACTION_PROFILE_ACTION_TYPE = "Can't create ActionProfile for MARC Bib record type with Create action";
 
   static final Map<String, String> ERROR_CODES_TYPES_RELATION = Map.of(
     "mappingProfile", "The field mapping profile",
@@ -516,7 +517,7 @@ public class DataImportProfilesImpl implements DataImportProfiles {
       try {
         entity.getProfile().setMetadata(getMetadata(okapiHeaders));
         composeFutureErrors(
-          validateProfile(OperationType.CREATE, entity.getProfile(), actionProfileService, tenantId),
+          validateActionProfile(OperationType.CREATE, entity.getProfile(), tenantId),
           validateActionProfileAddedRelationsFolioRecord(entity, tenantId)).onComplete(errors -> {
           if (errors.failed()) {
             logger.warn(format(PROFILE_VALIDATE_ERROR_MESSAGE, entity.getClass().getSimpleName()), errors.cause());
@@ -566,7 +567,7 @@ public class DataImportProfilesImpl implements DataImportProfiles {
           if (isDtoValidForUpdate) {
             entity.getProfile().setMetadata(getMetadata(okapiHeaders));
             return composeFutureErrors(
-              validateProfile(OperationType.UPDATE, entity.getProfile(), actionProfileService, tenantId),
+              validateActionProfile(OperationType.UPDATE, entity.getProfile(), tenantId),
               validateActionProfileChildProfilesFolioRecord(entity, tenantId, id),
               validateActionProfileAddedRelationsFolioRecord(entity, tenantId)
             ).compose(errors -> {
@@ -900,6 +901,17 @@ public class DataImportProfilesImpl implements DataImportProfiles {
         List<Error> fieldsValidationErrors = validateRepeatableFields(mappingProfile);
         errors.withTotalRecords(errors.getTotalRecords() + fieldsValidationErrors.size())
           .getErrors().addAll(fieldsValidationErrors);
+        return errors;
+      });
+  }
+
+  private Future<Errors> validateActionProfile(OperationType operationType, ActionProfile actionProfile, String tenantId) {
+    return validateProfile(operationType, actionProfile, actionProfileService, tenantId)
+      .map(errors -> {
+        if (ActionProfile.FolioRecord.MARC_BIBLIOGRAPHIC == actionProfile.getFolioRecord() && ActionProfile.Action.CREATE == actionProfile.getAction()) {
+          logger.warn("validateActionProfile:: {}", INVALID_ACTION_PROFILE_ACTION_TYPE);
+          errors.withTotalRecords(errors.getTotalRecords() + 1).getErrors().add(new Error().withMessage(INVALID_ACTION_PROFILE_ACTION_TYPE));
+        }
         return errors;
       });
   }