+++ mod-di-converter-storage-server/src/test/java/org/folio/rest/impl/ActionProfileTest.java
@@ -35,6 +35,7 @@ import static java.util.Collections.singletonList;
 import static org.folio.rest.impl.MappingProfileTest.MAPPING_PROFILES_PATH;
 import static org.folio.rest.jaxrs.model.ActionProfile.Action.CREATE;
 import static org.folio.rest.jaxrs.model.ActionProfile.Action.MODIFY;
+import static org.folio.rest.jaxrs.model.ActionProfile.Action.UPDATE;
 import static org.folio.rest.jaxrs.model.ActionProfile.FolioRecord.INSTANCE;
 import static org.folio.rest.jaxrs.model.ActionProfile.FolioRecord.MARC_BIBLIOGRAPHIC;
 import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.ACTION_PROFILE;
@@ -96,12 +97,12 @@ public class ActionProfileTest extends AbstractRestVerticleTest {
   static ActionProfileUpdateDto actionProfile_4 = new ActionProfileUpdateDto()
     .withProfile(new ActionProfile().withId(ACTION_PROFILE_UUID).withName("OLA")
       .withTags(new Tags().withTagList(Arrays.asList("lorem", "ipsum", "dolor")))
-      .withAction(CREATE)
+      .withAction(UPDATE)
       .withFolioRecord(MARC_BIBLIOGRAPHIC));
   static ActionProfileUpdateDto actionProfileNotEmptyChildAndParent = new ActionProfileUpdateDto()
     .withProfile(new ActionProfile()
       .withName("Action profile with child and parent")
-      .withAction(CREATE)
+      .withAction(UPDATE)
       .withParentProfiles(List.of(new ProfileSnapshotWrapper().withId(UUID.randomUUID().toString())))
       .withChildProfiles(List.of(new ProfileSnapshotWrapper().withId(UUID.randomUUID().toString())))
       .withFolioRecord(MARC_BIBLIOGRAPHIC));
@@ -177,15 +178,19 @@ public class ActionProfileTest extends AbstractRestVerticleTest {
   }
 
   @Test
-  public void shouldReturnBadRequestOnPost() {
-    createProfiles();
+  public void shouldReturnBadRequestOnPostWithMarcbibRecordAndCreateAction() {
     RestAssured.given()
       .spec(spec)
-      .body(new JsonObject().toString())
+      .body(new ActionProfileUpdateDto()
+        .withProfile(new ActionProfile()
+          .withName("Invalid Action Profile")
+          .withAction(CREATE)
+          .withFolioRecord(MARC_BIBLIOGRAPHIC)))
       .when()
       .post(ACTION_PROFILES_PATH)
       .then()
-      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY);
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY)
+      .body("errors[0].message", is("Can't create ActionProfile for MARC Bib record type with Create action"));
   }
 
   @Test
@@ -287,6 +292,26 @@ public class ActionProfileTest extends AbstractRestVerticleTest {
         .then()
         .statusCode(HttpStatus.SC_BAD_REQUEST);
     }
+  }
+
+  @Test
+  public void shouldReturnBadRequestOnPutWithMarcbibRecordAndCreateAction() {
+    RestAssured.given()
+      .spec(spec)
+      .body(actionProfile_4)
+      .when()
+      .post(ACTION_PROFILES_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_CREATED);
+
+    RestAssured.given()
+      .spec(spec)
+      .body(actionProfile_4.withProfile(actionProfile_4.getProfile().withAction(CREATE)))
+      .when()
+      .put(ACTION_PROFILES_PATH + "/" + actionProfile_4.getProfile().getId())
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY)
+      .body("errors[0].message", is("Can't create ActionProfile for MARC Bib record type with Create action"));
   }
 
   @Test
@@ -930,7 +955,7 @@ public class ActionProfileTest extends AbstractRestVerticleTest {
     var actionProfileUpdateDto = postActionProfile(new ActionProfileUpdateDto()
       .withProfile(new ActionProfile()
         .withName("Test Action Profile")
-        .withAction(CREATE)
+        .withAction(UPDATE)
         .withFolioRecord(MARC_BIBLIOGRAPHIC))
       .withAddedRelations(List.of(
         new ProfileAssociation()
@@ -943,7 +968,7 @@ public class ActionProfileTest extends AbstractRestVerticleTest {
       .body(new ActionProfileUpdateDto()
         .withProfile(new ActionProfile()
           .withName("Test Action Profile")
-          .withAction(CREATE)
+          .withAction(UPDATE)
           .withFolioRecord(MARC_BIBLIOGRAPHIC))
         .withAddedRelations(List.of(
           new ProfileAssociation()
@@ -1042,7 +1067,7 @@ public class ActionProfileTest extends AbstractRestVerticleTest {
     var actionProfileUpdateDto = postActionProfile(new ActionProfileUpdateDto()
       .withProfile(new ActionProfile()
         .withName("Test Action Profile")
-        .withAction(CREATE)
+        .withAction(UPDATE)
         .withFolioRecord(MARC_BIBLIOGRAPHIC)));
 
 
@@ -1062,7 +1087,7 @@ public class ActionProfileTest extends AbstractRestVerticleTest {
                                                                   Boolean expectedRemove9SubfieldFlag) {
     var actionProfile = new ActionProfileUpdateDto()
       .withProfile(new ActionProfile().withName("test:" + folioRecord)
-        .withAction(CREATE)
+        .withAction(UPDATE)
         .withFolioRecord(folioRecord)
         .withRemove9Subfields(incomingRemove9SubfieldFlag));