+++ mod-di-converter-storage-server/src/test/java/org/folio/rest/impl/MappingProfileTest.java
@@ -509,7 +509,7 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
       .spec(spec)
       .body(new ActionProfileUpdateDto().withProfile(new ActionProfile()
         .withName("testActionProfile")
-        .withAction(CREATE)
+        .withAction(UPDATE)
         .withFolioRecord(MARC_BIBLIOGRAPHIC)))
       .when()
       .post(ACTION_PROFILES_PATH);
@@ -930,7 +930,7 @@ public class MappingProfileTest extends AbstractRestVerticleTest {
     var actionProfileUpdateDto = postActionProfile(new ActionProfileUpdateDto()
       .withProfile(new ActionProfile()
         .withName("Test Action Profile")
-        .withAction(CREATE)
+        .withAction(UPDATE)
         .withFolioRecord(MARC_BIBLIOGRAPHIC)));
 
     RestAssured.given()